♻️ page임을 명확히 표기하기위해 네이밍 변경
- 네이밍 변경 및 기존 scss -> styled components로 변경
